Make thrusters appear for other players on clients
It was previously creating additional entities for the thrusters on the client that weren't parented to the player.
See https://docs.rs/bevy_replicon/latest/bevy_replicon/#component-relations

We only wanted one additional value to be sent over the network instead
of additional entities and transforms, so I put the thruster component
on the Player entity and had it manage updating its child thruster
visuals.